module3/company_api_with_orm/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy>=2.0.36
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
